fix(build-audit): move property input
